backend/movie/pyproject.toml
[project]
name = "movie"
version = "0.1.0"
description = "MovieService"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=2.21.1",
    "aiokafka>=0.12.0",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "fastapi>=0.115.11",
    "loguru>=0.7.3",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.5",
    "python-magic>=0.4.27",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.39",
    "strawberry-graphql>=0.262.3",
    "uvicorn>=0.34.0",
]
